Using `mrb_realloc()` with `stack_extend_alloc()`
Raising a `SystemStackError` exception in an out-of-memory situation will generate backtrace information.
This can eventually lead to a `NoMemoryError` exception, and this process flow is completely undesirable.

In effect, this change means that it will include a revert of commit 0dbb9e6e411572b0c37e7be723d6dc3895387f96.