Freezing objects by the "asz" specifier of `mrb_get_args()`
The `a`, `s`, and `z` specifiers of `mrb_get_args()` allow you to get the address for an element of an object.

If the caller of `mrb_get_args()` subsequently calls `mrb_vm_exec()`, an operation may be performed that changes the address obtained.
When the caller references the changed address in a subsequent operation, use-after-free is established.

Since this patch makes the address immutable, it will copy the object and freeze it if necessary.
Note that the price of safety is a reduction in both memory usage and processing speed.
If this is a concern, users are encouraged to use the `A` and `S` specifiers.