Wrap `mrb_raise()` and `mrb_raisef()` with macros
This change is mainly to avoid use-after-free, which tends to occur when C strings are passed as arguments.

Here is a scenario in which use-after-free is valid:

  - Remove the `RuntimeError` class constant.
  - Define the `Object.const_missing` method.
  - The address to the C string is taken from the string object by `mrb_get_args()` and so on.
  - To resolve `E_RUNTIME_ERROR`, `Object.const_missing` is called via `mrb_exc_get_id()`.
  - Make a major change to the string object to invalidate the address of the C string.
  - The `mrb_str_new_cstr()` called by `mrb_raise()` references the invalidated C string.

This patch is reimplemented with macros to ensure that the string object is created and then `mrb_exc_get_id()` is called.

For compatibility when linked as shared object files or language bindings, the original function symbols are still available.

**INCOMPATIBILITY NOTE**:
These macro functions are now statements rather than expressions.
So they can no longer be used inside the ternary and comma operators.