Implement `to_hash` calling in implicit conversion contexts
Previously mruby did not call `#to_hash` on objects in contexts like
the keyword splat operator. This patch implements that feature for
feature parity with CRuby. Additionally a call to `mrb_ensure_hash_type`
in the implementation of `OP_HASHADD` has been changed to an `mrb_assert`
as the only situation in which it could trigger would be if the bytecode
itself contradicted its assumptions (a similar `assert` was already present
in the implementation of `OP_HASHCAT`).